move options to options.go
