+++ placement/options.go
@@ -1,0 +1,79 @@
+// Copyright (c) 2016 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package placement
+
+const defaultMaxStepSize = 3
+
+type deploymentOptions struct {
+	maxStepSize int
+}
+
+// NewDeploymentOptions returns a default DeploymentOptions
+func NewDeploymentOptions() DeploymentOptions {
+	return deploymentOptions{maxStepSize: defaultMaxStepSize}
+}
+
+func (o deploymentOptions) MaxStepSize() int {
+	return o.maxStepSize
+}
+
+func (o deploymentOptions) SetMaxStepSize(stepSize int) DeploymentOptions {
+	o.maxStepSize = stepSize
+	return o
+}
+
+// NewOptions returns an Options instance
+func NewOptions() Options {
+	return options{}
+}
+
+type options struct {
+	looseRackCheck      bool
+	acrossZones         bool
+	allowPartialReplace bool
+}
+
+func (o options) LooseRackCheck() bool {
+	return o.looseRackCheck
+}
+
+func (o options) SetLooseRackCheck(looseRackCheck bool) Options {
+	o.looseRackCheck = looseRackCheck
+	return o
+}
+
+func (o options) AcrossZones() bool {
+	return o.acrossZones
+}
+
+func (o options) SetAcrossZones(acrossZones bool) Options {
+	o.acrossZones = acrossZones
+	return o
+}
+
+func (o options) AllowPartialReplace() bool {
+	return o.allowPartialReplace
+}
+
+func (o options) SetAllowPartialReplace(allowPartialReplace bool) Options {
+	o.allowPartialReplace = allowPartialReplace
+	return o
+}
