+++ placement/placement.go
@@ -334,41 +334,3 @@ func (h host) Weight() uint32 {
 func (h host) String() string {
 	return fmt.Sprintf("[id:%s, rack:%s, zone:%s, weight:%v]", h.id, h.rack, h.zone, h.weight)
 }
-
-// NewOptions returns an Options instance
-func NewOptions() Options {
-	return options{}
-}
-
-type options struct {
-	looseRackCheck      bool
-	acrossZones         bool
-	allowPartialReplace bool
-}
-
-func (o options) LooseRackCheck() bool {
-	return o.looseRackCheck
-}
-
-func (o options) SetLooseRackCheck(looseRackCheck bool) Options {
-	o.looseRackCheck = looseRackCheck
-	return o
-}
-
-func (o options) AcrossZones() bool {
-	return o.acrossZones
-}
-
-func (o options) SetAcrossZones(acrossZones bool) Options {
-	o.acrossZones = acrossZones
-	return o
-}
-
-func (o options) AllowPartialReplace() bool {
-	return o.allowPartialReplace
-}
-
-func (o options) SetAllowPartialReplace(allowPartialReplace bool) Options {
-	o.allowPartialReplace = allowPartialReplace
-	return o
-}
